// src/shared/api/generated/model/issueOtpRequest.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * CoreBank API Specification
 * Corebank REST API 명세서
 * OpenAPI spec version: v1.0.0
 */
import type { IssueOtpRequestTransactionData } from './issueOtpRequestTransactionData';
import type { IssueOtpRequestTransactionType } from './issueOtpRequestTransactionType';

export interface IssueOtpRequest {
  transactionType: IssueOtpRequestTransactionType;
  transactionData: IssueOtpRequestTransactionData;
}