// src/shared/api/generated/model/verifyOtpRequest.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * CoreBank API Specification
 * Corebank REST API 명세서
 * OpenAPI spec version: v1.0.0
 */

export interface VerifyOtpRequest {
  /** @minLength 1 */
  otpRequestId: string;
  /**
     * @minLength 1
     * @pattern ^\d{6}$
     */
  otpCode: string;
}